# jsonplaceholder.openapi.yaml
openapi: 3.0.3
info:
  title: JSON Placeholder
  description: JSONPlaceholder is a free online REST API that you can use whenever you need some fake data. It can be in a README on GitHub, for a demo on CodeSandbox, in code examples on Stack Overflow, ...or simply to test things locally
  version: 0.0.1
servers:
  - url: 'https://jsonplaceholder.typicode.com/'
paths:
  /posts:
    get:
      tags: [ post ]
      operationId: posts
      summary: Listing all resources
      parameters:
        - in: query
          name: userId
          description: Filter - this will return all the posts that belong to the user with specified id
          schema: { $ref: '#/components/schemas/PositiveInt' }
      responses:
        200:
          description: A JSON array of posts
          content: { application/json: { schema: { $ref: '#/components/schemas/Posts' } } }
    post:
      tags: [ post ]
      operationId: addPost
      summary: Create a post
      description: Important - resource will not be really updated on the server but it will be faked as if
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Post' }
            example:
              title: 'foo'
              body: 'bar'
              userId: 1
      responses:
        201:
          description: Post will not be really created on the server but it will be faked as if
          content: { application/json: { schema: { $ref: '#/components/schemas/Post' } } }
  /posts/{id}:
    get:
      tags: [ post ]
      operationId: pickPost
      summary: Get Post by id
      parameters:
        - name: id
          in: path
          required: true
          description: id of post in a post list
          schema: { $ref: '#/components/schemas/PositiveInt' }
      responses:
        200:
          description: JSON of Post entity, specified by it's id
          content: { application/json: { schema: { $ref: '#/components/schemas/Post' } } }
        404: { $ref: '#/components/responses/404' }
    put:
      tags: [ post ]
      operationId: updatePost
      summary: Updating a resource
      description: Important - resource will not be really updated on the server but it will be faked as if
      parameters:
        - name: id
          in: path
          required: true
          description: id of post in a post list
          schema: { $ref: '#/components/schemas/PositiveInt' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Post' }
            example:
              id: 1
              title: 'foo'
              body: 'bar'
              userId: 1
      responses:
        204:
          description: The server successfully processed the request, but is not returning any content
    patch:
      tags: [ post ]
      operationId: mergePost
      summary: Patching a resource
      parameters:
        - name: id
          in: path
          required: true
          description: id of post in a post list
          schema: { $ref: '#/components/schemas/PositiveInt' }
      responses:
        200:
          description: The server successfully processed the request, but is not returning any content
          content: { application/javascript: { schema: { $ref: '#/components/schemas/Post' } } }
    delete:
      tags: [ post ]
      operationId: removePost
      summary: Deleting a resource
      parameters:
        - name: id
          in: path
          required: true
          description: id of post in a post list
          schema: { $ref: '#/components/schemas/PositiveInt' }
      responses:
        200:
          description: resource will not be really updated on the server but it will be faked as if
          content: { application/json: { schema: { $ref: '#/components/schemas/EmptyObject' } } }
  /posts/{id}/comments:
    get:
      tags: [ post, comment ]
      operationId: postComments
      summary: Listing all the comments of post
      description: One level of nested route is available
      parameters:
        - name: id
          in: path
          required: true
          description: id of post in a post list
          schema: { $ref: '#/components/schemas/PositiveInt' }
      responses:
        200:
          description: A JSON array of comments
          content: { application/json: { schema: { $ref: '#/components/schemas/Comments' } } }
        404: { $ref: '#/components/responses/404' }
    post:
      tags: [ post, comment ]
      operationId: addCommentToPost
      summary: Create a comment
      description: Important - resource will not be really updated on the server but it will be faked as if
      parameters:
        - name: id
          in: path
          required: true
          description: id of post in a post list
          schema: { $ref: '#/components/schemas/PositiveInt' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Comment' }
            example:
              postId: 1
              name: 'foo'
              email: 'baz'
              body: 'bar'
      responses:
        201:
          description: Post will not be really created on the server but it will be faked as if
          content: { application/json: { schema: { $ref: '#/components/schemas/Comment' } } }
  /comments:
    get:
      tags: [ comment ]
      operationId: comments
      summary: Listing all comments
      description: All the comments for all entities presented in this API. Can be filtered by optional query params
      parameters:
        - in: query
          name: postId
          schema: { $ref: '#/components/schemas/PositiveInt' }
          description: Filter - this will return all the comments that belong to the post with specified id
      responses:
        200:
          description: A JSON array of comments
          content: { application/json: { schema: { $ref: '#/components/schemas/Comments' } } }
    post:
      tags: [ comment ]
      operationId: addComment
      summary: Create a comment
      description: Important - resource will not be really updated on the server but it will be faked as if
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Comment' }
            example:
              postId: 1
              name: 'foo'
              email: 'baz'
              body: 'bar'
      responses:
        201:
          description: Comment will not be really created on the server but it will be faked as if
          content: { application/json: { schema: { $ref: '#/components/schemas/Comment' } } }
...
components:
  responses:
    404:
      description: No Content - empty json array
      content: { application/json: { schema: { $ref: '#/components/schemas/EmptyObject' } } }
  schemas:
    PositiveInt:
      type: integer
      format: int32
      minimum: 1
    EmptyObject:
      type: object
    Post:
      type: object
      properties:
        id: { $ref: '#/components/schemas/PositiveInt' }
        userId: { $ref: '#/components/schemas/PositiveInt' }
        title: { type: string }
        body: { type: string }
    Posts:
      type: array
      items: { $ref: '#/components/schemas/Post' }
    Comment:
      description: Comment for post
      type: object
      properties:
        id: { $ref: '#/components/schemas/PositiveInt' }
        postId: { $ref: '#/components/schemas/PositiveInt' }
        name: { type: string }
        email: { type: string }
        body: { type: string }
    Comments:
      type: array
      items: { $ref: '#/components/schemas/Comment' }
...